requirements.txt
annotated-types==0.7.0
babel==2.17.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
courlan==1.3.2
dateparser==1.2.2
et_xmlfile==2.0.0
htmldate==1.9.4
idna==3.11
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
numpy==2.4.0
openpyxl==3.1.5
pandas==2.3.3
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
six==1.17.0
soupsieve==2.8.1
tld==0.13.1
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
